feat: Update schema to make category_id optional on products
This update makes the `category_id` on the `products` to be optional so
that it doesn't need to be provided when a product does not fit into a
category yet.

It also ensures that when a category is deleted, the products that
depend will have their `category_id` field set to null.

Finally, `available` and `currency` have been introduced to the
`products` to table to help know the availability status of a product
and the currency it uses.