feat: Add background jobs for order management
This update includes relevant background jobs for handling orders.
Immediately a new order is create (order is checked out), a
notification is sent to the payment service to start processing it.

It continues to track the status until the order is complete(from the
payment service) or cancelled by the user.