feat: Refactor order filtering and caching
This commit refactors the order filtering and caching logic in the `Api::OrdersController`. The unnecessary `latest_update_time` variable has been removed, as it was not being used. The `last_update` filter has also been removed from the `cache_collection` method call. These changes improve the efficiency and readability of the code.