Refactor and Update: Order Filtering, Caching, Authentication, Jobs, and API Enhancements (#14)
* feat: Refactor order filtering and caching

This commit refactors the order filtering and caching logic in the `Api::OrdersController`. The unnecessary `latest_update_time` variable has been removed, as it was not being used. The `last_update` filter has also been removed from the `cache_collection` method call. These changes improve the efficiency and readability of the code.

* feat: Refactor authentication.rb for improved readability and parameter naming

* feat: Refactor UserServiceClient to use new API endpoints

This commit updates the `UserServiceClient` class to use the new API endpoints for validating developer tokens, users, and apps. It also improves error handling and caching logic. These changes enhance the reliability and performance of the user service client.

* feat: Add 'refunded' status to order

This commit adds the 'refunded' status to the 'order_status' enum in the database schema. It allows orders to be marked as refunded, providing more flexibility in tracking order statuses.

* feat: Add app_id parameter to fetch_developer_id and fetch_user methods

This commit adds the `app_id` parameter to the `fetch_developer_id` and `fetch_user` methods in the `UserServiceClient` class. This allows for more specific fetching of developer IDs and users based on the associated app. The changes improve the flexibility and accuracy of the user service client.

* chore: Update rails_helper.rb to set MESSAGE_BROKER_URL environment variable

This commit updates the `rails_helper.rb` file to set the `MESSAGE_BROKER_URL` environment variable to `redis://localhost:6379/1`. This change ensures that the correct message broker URL is used for the test environment, improving the reliability and accuracy of the tests.

* feat: Add ascending order functionality to order API endpoint

* chore: Update database configuration for the application

* test: Update tests for background jobs

* feat: Add 'refunded' status to order

* Refactor NotifyUserServiceJob to handle order not found error

This commit refactors the `NotifyUserServiceJob` class to handle the case when the order with the given ID is not found. Instead of using a negative condition (`unless order`), it now uses a positive condition (`if order.nil?`) to check if the order is nil. If the order is not found, an error message is logged and the method returns. This change improves the readability and maintainability of the code.

* feat: Update PaymentStatusJob to handle 'refunded' status

This commit updates the `PaymentStatusJob` class to handle the 'refunded' status for orders. It adds the logic to update the order status to 'refunded' and perform necessary actions such as notifying the user and updating the product stock. Additionally, it introduces a private method `update_order_and_notify` to encapsulate the code for updating the order status and notifying the user. This change improves the functionality and flexibility of the payment status handling in the application.

* feat: Update UpdateProductStockJob to handle 'refunded' status

This commit updates the `UpdateProductStockJob` class to handle the 'refunded' status for orders. It adds the logic to update the product stock based on the refunded quantity. The stock quantity is increased by the refunded quantity for the respective product. This change improves the accuracy of the stock management system in the application.

* feat: Update bin scripts to handle long lines in bundle binstub

This commit updates the bin scripts (`htmldiff`, `ruby-parse`, `ruby-rewrite`) to handle long lines in the bundle binstub file. It modifies the code to properly handle long lines by splitting them into multiple lines for improved readability. This change ensures that the bundle binstub file can be processed correctly, preventing any potential issues related to long lines.